Validate EC chain invariants on instantiation (#138)
EC chain cannot contain a zero-valued tip set. Enforce this in the
exported API and return an error when such tip set is specified in the
chain.

Add tests to assert the EC chain invariants are met:
* Epochs in chain are weakly and monotonically increasing, where
repeated epochs are not allowed but gaps may exist in the epochs.
* Base cannot be a zero-valued tip set, as defined by `TipSet.IsZero`.

Implement `ECChain.Validate` to encapsulate the chain validation logic,
and validate all received chains by a `Participant`. Further, strictly
refuse to receive zero-valued chains, since a zero-valued chain is valid
but not acceptable by by a `Participant`. The `ECChain` constructor is
also refactored to reuse the newly encapsulated validation logic.

Additionally, optimise check for base, prefix, and equality based on the
invariants listed above. Specifically, avoid reflection when checking
for equality for better performance.

Fixes #128