Refactor `Host` interface to optionally trace logic logs (#140)
Tracing logic logs is not a fundamental function with gpbft algorithm.
Instead, it is a crosscutting concern that currently aids one to reason
about the logical changes of a participant.

Refactor the core `Host` interface to relax the requirement for
implementing the logic log tracing functionality. Introduce a separate
`Tracer` interface that when specified logs logic change logs.

This change separates the tracing requirements introduced primarily by
the `sim` package from the core gpbft algorithm interfaces.
Additionally, it makes mocking the `gpbft` behaviour for testing
purposes simpler.

Note that the tracing concept in simulation defers from the
"system logging", in that: it is a configurable textual output of a
simulation. Therefore, the work here keeps them separate instead of
replacing it with a typical logging library.