Consistently model P2P latency to reduce max rounds in tests (#212)
The latency model in simulations is generated using a Log Normal
distribution with some configured mean. The latency generation occurs
every time a message is sent from any one participant to another
participant. This effectively simulates an adhoc network with volatile
latency which is typically observed in mobile networks.

Compared to Filecoin network, this approach is unrealistic and too
volatile. Because, in Filecoin SPs are typically well-connected servers
with high bandwidth low latency connection to their peers. As a result,
tests regularly exceed the maximum number of rounds where, so far, they
require at least 21 rounds to succeed.

A more realistic latency model is one that uses some randomly generated
but fixed point-to-point latency that remains constant or varies very
slightly over the lifetime of a participant. The changes here introduce
consistent P2P modeling of latency for Log Normal latency model, where
the latency from one participant to another is generated once and
remains constant for the lifetime of a simulation. The latency model
uses an asymmetric model where, latency from one node to another is not
necessarily the same other way around.

This change reduces the maximum number of rounds required
for tests from 30 to 10.

Fixes #196