Drop spammable messages beyond max allowed lookahead rounds (#258)
The gPBFT rebroadcast design revision introduced an improvement whereby
"spammable" messages, i.e. `COMMIT` for bottom, for future rounds should
be dropped by the participants. The specification leaves room for
implementers to optionally retain some number of such messages to
reduce reliance on the need for rebroadcast and ultimately help reach
consensus in fewer rounds.

The work here:
1. introduces `WithMaxLookaheadRounds` configuration option to `gpbft`
   package, defaulting to zero if unset.
2. drops all spammable messages belonging to rounds that are beyond the
   configured max lookahead rounds.

A `Spam` adversary is introduced to replicate the conditions at which
max lookahead rounds would start dropping messages. A set of tests then
use the `Spam` adversary to assert that honest participants reach
expected consensus despite the presence of spam messages.

Note that there are no APIs to assert that the upper lookahead bound is
respected by the implementation. A TODO is left to revisit this once
telemetry is introduced whereby tests can observe future message queue
size metrics to assert a configured lookahead is indeed respected.

Resolves #240