Allow skipping to future rounds from QUALITY phase (#279)
Update the core protocol implementation to allow skipping to future
rounds from QUALITY phase. This reduces the conditions to requiring an
instance to only fully execute DECIDE phase. Otherwise, the instance may
jump ahead as long as all queued messages are processed.

As a result, the implementation no longer needs to check for QUALITY
phase transition on receiving alarms.

Fixes #278