Verify decider identity when checking if sim instance is complete (#323)
Simulation checks if an instance is complete (i.e. expected participants
has reached some decision) at every network tick. Prior to the changes
here this was determined using the number of decisions received.

A problem with this approach is that when some adversary targets are
excluded from consensus check a simulation instance may deemed to be
complete despite non-targeted participants not having reached a decision
yet.

The changes here explicitly checks the ID of participants to assure that
only decisions from intended participants are counted towards the total
number of expected decisions.