Add basic certificate exchange metrics (#529)
* Add basic certificate exchange metrics

This adds basic metrics so we can know how many requests we're making,
how many peers we're talking to, how often we're polling, etc.

I still want to add some metrics around how often we're polling when we
don't _need_ to, but I also want to rework some of that logic at the
same time to reduce the number of unnecessary polls. So I'm punting that
out of this round of work.

* address feedback

* rename polling meter

* refactor server/client metrics

* Refactor subscriber metrics

* And the leftover bugs...

* remove units

* prefix everything with "polling_"