Rebroadcast already signed messages from WAL (#691)
* Rebroadcast already signed messages from WAL

Previously, rebroadcast functioned by keeping track of message templates
and calling `RequestBroadcast` again to re-broadcast them. This approach
can cause erroneous behaviour when messages are re-signed in conjunction
with participant churn.

To avoid this category of errors treat broadcast final: Every
broadcasted message is recorded in WAL and once recorded the same
message is used for re-broadcast; no re-signing. A dedicated
re-broadcast API is introduced to take instance, round and phase and use
it to then look up broadcast messages.

Note that as a result of the changes above, failures are also sicky. In
that, if signing of a message fails at the time of original broadcast,
future requests to rebroadcast it will have no effect.

The changes here also include a series of WAL functionality refactors to
make it easier to traverse write-ahead log entries.

Fixes #474

* Keep wal exactly as dumb as it is and index messages externally