Introduce `gpbft.Instant` to capture precise progress moment (#698)
In a number of places, namely rebroadcast request and progress we need
information a precise moment in GPBFT progress. Prior to changes here
the changes were communicated as three separate values, which was
cumbersome to work with as the usage of this information increased with
rebroadcast.

Introduce `Instant` as a struct to capture precise instant in GPBFT
progress and reflect changes across `Progress` as well as
`RequestRebroadcast`.

Relates to https://github.com/filecoin-project/go-f3/pull/691#discussion_r1786665136