Use the latest computed power table instead of always ranging certs (#882)
Certificate store already keeps track of the latest power table, i.e.,
the power table of the next expected certificate. Use this pre-computed
table whenever the next power table is asked for instead of re-computing
it every time.

This results in a much faster response to `GetPowerTable` for the most
common case: getting the committee as part of monotonic progress of
GPBFT instances.

Fixes #881