Bump version in prep for release (#883)
Considering F3 has been running on butterflynet for a while and to
unblock the Lotus bubble-up PR cutting a release pre-emptively.

There will most likely be further go-f3 releases prior to calibnet and
mainnet upgrades.

Bumped the version straight to `v0.8.0` considering the breadth of
changes in chain exchange, and partial message propagation.