Parameterise F3 runner pusub constants into manifest (#903)
Parameterise the buffer size for the gmessage subscription passed to
pubsub, and the channel of validated messages to which messages are
written for processing by gpbft.

This is to cover all the corners in prep for mainnet passive testing,
since in the past we have had to bump thes values due to slow
processing of messages in specific cases.