Implement lag awareness in aider based on current progress (#992)
* Implement lag awareness in aider based on current progress

Determine if F3 is lagging and if so engage rebroadcast aid with
resilience against failure cases. The lag determination strategy uses
both lotus daemon and observer as sources to determine the current F3
progress. It then decides if F3 is lagging behind based on:
 * number of rounds
 * distance from head
 * elapsed time since instance ID changed.
 * last successful lag determination

The combination of above assures that both in cases of failure or less
than perfect information aid is attempted. The only case where aid can't
be attempted is when no initial F3 state is determinable.

Fixes #988

* Make minimum round confiurable