Infer messages from evidence of justification (#1024)
* Infer messages from evidence of justification

The justification accompanied by most messages in GPBFT is evidence of
a strong quorum of some states across the participants. Treat the
receipt of such justifications as the receipt of the messages that
justified it.

This allows a faster progress towards quorum in the network, and reduces
reliance on re-broadcast and majority network state propagation.

Specifically, there are 2 phases that can benefit from the
above optimization:

1. PREPARE phase, using justification of PREPARE in:
  * COMMIT messages at the same round.
  * PREPARE messages at the next round.
  * CONVERGE messages at the next round.
2. COMMIT phase, using justification of COMMIT to bottom in:
  * CONVERGE messages at the next round.
  * PREPARE messages at the next round.

The above optimization doesn't matter in CONVERGE and DECIDE phases.
Because, CONVERGE by design awaits the phase timeout by design, and
DECIDE is already optimized by `skipToDecide` where the participant
rightly so immediately progresses to the final phase using the
justification of COMMIT accompanied by the DECIDE message.

Fixes #342

* Add tests and fix logic flow in evaluating current round for COMMIT

Change the message processing logic to proceed to evaluate the current
round if a COMMIT message could make a difference to PREPARE completion.

Add emulator tests to assert PREPARE and COMMIT phases complete
immediately if the right justification arrives.