Fix bug in F3 stop and use start/stop in catchup tests (#1029)
Fix a bug where F3.Stop was never calling `stopInternal` which actually
resets the state and stops the internal components.

Remove the testing API exposed for pause resume in favor of stop start.
Because, now that manifest is final, there's no pause/resume and the
scenario that was being tested is irrelevant. Instead, stop the nodes
completely then start them back up to assert that they do catch up.

The flaky test was repeated for 10 minutes on CI without failing.

Fixes: #1025